feat: making torchserve usable without torch
