+++ docs/api/ts.protocol.rst
@@ -12,6 +12,14 @@ ts.protocol.otf\_message\_handler module
    :undoc-members:
    :show-inheritance:
 
+ts.protocol.otf\_torch\_message\_handler module
+----------------------------------------
+
+.. automodule:: ts.protocol.otf_torch_message_handler
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 
 Module contents
 ---------------
